Update: Sync whole dist folder per default.
This is needed for apps that rely on ember-cli-addons that
bundle their own assets. These addons will build their assets
to `dist/<addon-name>` thus we need to sync the whole dist folder
to make this work with ember-deploy.

This fixes #12. Also this resolves #7 because we get the static
hosted strategy for free in a rudimentary way due to the fact that
`index.html` will also be synced to s3 with the s3-adapter.