Use ember-deploy adapters for index/list/activate.
We can now remove the dependency on the deployinator project
because handling the key/value-store updates is handled by
the index-adapter layer since the last commits.

I also added the notion of revision taggers. Which in the future
should also be pluggable. Currently there is only one sort of
revisionTagger and that's the `sha` tagging adapter.

* remove deployinator dependency.
* add revisionTagger.
* refactor index/list/activate tasks.