Adds support for “plugin packs”.
A plugin pack is an addon with the keyword “ember-cli-deploy-plugin-pack”
and one or more dependent addons that are ember-cli-deploy-plugins.

Note that the plugin pack’s dependent addons should be listed as
dependencies in the pack’s package.json, not as devDependencies.

I expect that plugin packs may also implement a config/deploy.js
blueprint that is auto-executed upon `ember install` of the pack.