Fix plugin context merging in the pipeline model
Since upgrading to lodash `^4.0.0` in v0.6.2 pipeline context merging
has been broken. When your hook result has an object that exists in
the context already (e.g. `distFiles`) then the results of your hook are
no longer concatenated but instead overwrite the existing values.

For example, if you have `context.distFiles` set to `['file1', file2']`
and your hook returns `{distFiles: ['file3']}` then the resulting array
is now `['file3', 'file2']`, not `['file1', 'file2', 'file3']` as it was
before.

This is because the `_.merge` function used to merge the contexts was
changed in lodash 4 and is now a simple merger that takes an object and
one or more sources to merge into that object; it no longer takes a custom
merge function. To use a merge function you need to change to use
`_.mergeWith`.

I've made this change and added a test to ensure regression does not
happen in future.