[DOC] Update Gemfile.lock for docs
The gemfile uses a dynamic version from the github api. I was not able
to `bundle install` without updating Gemfile.lock since the dependencies
needed for the latest version of `github-pages` has since changed. It
looks like the last time this was updated was 6 months ago, so it makes
sense this will need to be updated.